+++ index.html
@@ -105,7 +105,7 @@
         </fieldset>
         <fieldset class="ad-form__element">
           <label class="ad-form__label" for="price">Цена за ночь, руб.</label>
-          <input id="price" name="price" type="number" placeholder="5000" required max="100000" data-pristine-required-message="Обязательное поле" data-pristine-max-message="Максимальное значение - 100000">
+          <input id="price" name="price" type="number" placeholder="1000" required min="0" max="100000" data-pristine-required-message="Обязательное поле" data-pristine-min-message="Значение не должно быть отрицательным" data-pristine-max-message="Максимальное значение - 100000">
           <div class="ad-form__slider"></div>
         </fieldset>
         <fieldset class="ad-form__element ad-form__element--time">
