+++ js/main.js
@@ -1,6 +1,7 @@
 import './util.js';
 import './popup.js';
 import './form-setup.js';
+import './form-validation.js';
 import { createAdverts } from './data.js';
 
 createAdverts();
